requirements.txt
crossplane==0.5.8
redis==5.0.0
Flask-Session==0.5.0
mkdocs-mermaid2-plugin

